pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-gateway"
version = "1.0.0"
description = "Universal MCP Gateway - Single-port multiplexing with Meta-MCP for 95% token savings"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Mikko Parkkola", email = "mikko@parkkola.fi" }
]
keywords = [
    "mcp",
    "model-context-protocol",
    "ai",
    "llm",
    "gateway",
    "proxy",
    "multiplexer",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Typing :: Typed",
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "pytest-timeout>=2.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
    "httpx>=0.27",  # For testing
]
metrics = [
    "prometheus-client>=0.20",
]
all = ["mcp-gateway[dev,metrics]"]

[project.scripts]
mcp-gateway = "mcp_gateway.cli:main"

[project.urls]
Homepage = "https://github.com/MikkoParkkola/mcp-gateway"
Documentation = "https://github.com/MikkoParkkola/mcp-gateway#readme"
Repository = "https://github.com/MikkoParkkola/mcp-gateway"
Issues = "https://github.com/MikkoParkkola/mcp-gateway/issues"
Changelog = "https://github.com/MikkoParkkola/mcp-gateway/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_gateway"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/examples",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
]

[tool.ruff.lint.isort]
known-first-party = ["mcp_gateway"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["aiohttp.*", "yaml.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-x",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["src/mcp_gateway"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
